- Fix 255 limit of byte variable for next pose - Turn off all debug information to test the maximum speed.

+++ src/V2_Command.h
@@ -7,7 +7,7 @@
 bool V2_ActionPlaying = false;
 byte V2_ActionCombo = 0; // combo = 0 reserved for single action play
 byte V2_NextAction = 0;
-byte V2_NextPose = 0;
+uint16_t V2_NextPose = 0;
 unsigned long V2_NextPlayMs = 0;
 int V2_ServoTimeRatio = 20;
